# api/openapi.yaml
openapi: 3.0.3
info:
  title: Agora API
  version: 1.0.0
  description: |
    Agora School Management Platform API.
    Base contract aligned with Step 3 (auth, attendance, homework, marks, messaging, notifications).
servers:
  - url: /api/v1
    description: Relative base URL
...
paths:
...
  /conversations:
    get:
      tags: [Messaging]
      operationId: listConversations
      summary: List user conversations
      parameters:
        - $ref: '#/components/parameters/pageQuery'
        - $ref: '#/components/parameters/pageSizeQuery'
      responses:
        '200':
          description: Conversation list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Messaging]
      operationId: createConversation
      summary: Create conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationCreateRequest'
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /conversations/{conversationId}/messages:
    get:
      tags: [Messaging]
      operationId: listConversationMessages
      summary: List messages in a conversation
      parameters:
        - $ref: '#/components/parameters/conversationIdPath'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        '200':
          description: Message list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags: [Messaging]
      operationId: createMessage
      summary: Send message to conversation
      parameters:
        - $ref: '#/components/parameters/conversationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreateRequest'
      responses:
        '201':
          description: Message created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /conversations/{conversationId}/read:
    post:
      tags: [Messaging]
      operationId: markConversationRead
      summary: Mark conversation as read for current user
      parameters:
        - $ref: '#/components/parameters/conversationIdPath'
      responses:
        '200':
          description: Read state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessBooleanResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceApiKey:
      type: apiKey
      in: header
      name: X-Device-Api-Key
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-Api-Key
  parameters:
    pageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSizeQuery:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    conversationIdPath:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
    PagingMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
    ErrorItem:
      type: object
      properties:
        field:
          type: string
        issue:
          type: string
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/Meta'
...
    MessageKind:
      type: string
      enum: [text, file, system]
...
    SuccessBooleanResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            ok:
              type: boolean
              example: true
        meta:
          $ref: '#/components/schemas/Meta'
...
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [direct, group, broadcast]
        title:
          type: string
          nullable: true
        unread_count:
          type: integer
        last_message_preview:
          type: string
          nullable: true
    ConversationCreateRequest:
      type: object
      required: [kind, participant_user_ids]
      properties:
        kind:
          type: string
          enum: [direct, group, broadcast]
        title:
          type: string
          nullable: true
        participant_user_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    ConversationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Conversation'
        meta:
          $ref: '#/components/schemas/Meta'
    ConversationListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
            - type: object
              properties:
                pagination:
                  $ref: '#/components/schemas/PagingMeta'
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        sender_user_id:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/MessageKind'
        body:
          type: string
          nullable: true
        attachment_urls:
          type: array
          items:
            type: string
            format: uri
        sent_at:
          type: string
          format: date-time
        edited_at:
          type: string
          format: date-time
          nullable: true
    MessageCreateRequest:
      type: object
      required: [kind]
      properties:
        kind:
          $ref: '#/components/schemas/MessageKind'
        body:
          type: string
        attachment_urls:
          type: array
          items:
            type: string
            format: uri
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Message'
        meta:
          $ref: '#/components/schemas/Meta'
    MessageListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
            - type: object
              properties:
                next_cursor:
                  type: string
                  nullable: true
...
security:
  - BearerAuth: []